Created an intermediate script to handle a "lite" version of BAC0 without pandas, flask and bokeh Modified filepath of log file now in the user space under .BAC0 removed to_excel function A lot of tweaks to handle the lack of pandas in the history output or in the devices property
