+++ BAC0/infos.py
@@ -12,5 +12,5 @@ __author__ = 'Christian Tremblay, P.Eng.'
 __email__ = 'christian.tremblay@servisys.com'
 __url__ = 'https://github.com/ChristianTremblay/BAC0'
 __download_url__ = 'https://github.com/ChristianTremblay/BAC0/archive/master.zip'
-__version__ = '0.99.900'
+__version__ = '0.99.902'
 __license__ = 'LGPLv3'
