+++ BAC0/scripts/__init__.py
@@ -4,5 +4,9 @@
 # Copyright (C) 2015 by Christian Tremblay, P.Eng <christian.tremblay@servisys.com>
 # Licensed under LGPLv3, see file LICENSE in this source tree.
 #
-from . import BasicScript
-from . import ReadWriteScript
+from . import Base
+from . import Lite
+try:
+    from . import Complete
+except ImportError:
+    _PANDAS = False
